refactor utils and useEffect
