+++ app/packages/core/src/components/Actions/Selected.tsx
@@ -14,12 +14,16 @@ import * as fos from "@fiftyone/state";
 import { State } from "@fiftyone/state";
 import { ActionOption } from "./Common";
 import Popout from "./Popout";
+import { isImageSimilaritySearch } from "./similar/Similar";
 
 const useClearSampleSelection = (close) => {
   return useRecoilTransaction_UNSTABLE(
     ({ set }) =>
       async () => {
         set(fos.selectedSamples, new Set());
+        if (!fos.similarityParameters) {
+          set(isImageSimilaritySearch, false);
+        }
         close();
       },
     [close]
@@ -151,7 +155,7 @@ const useModalActions = (
 ) => {
   const selected = useRecoilValue(fos.selectedSamples);
   const clearSelection = useClearSampleSelection(close);
-
+  const hasSorting = useRecoilValue(fos.similarityParameters);
   const selectedLabels = useRecoilValue(fos.selectedLabelIds);
   const visibleSampleLabels = lookerRef.current
     ? lookerRef.current.getCurrentSampleLabels()
